refactor: 불필요한 소스파일 제거 및 디렉토리 이동
기능 목록서 재정의 하기 전에 남아있던 소스파일 제거
쓸모있는 클래스들은 디렉토리 이동
