+++ src/test/java/racingcar/domain/game/mock/CanMoveNumberGenerator.java
@@ -1,10 +1,10 @@
-package racingcar.player.mock;
+package racingcar.domain.game.mock;
 
 import static camp.nextstep.edu.missionutils.Randoms.pickNumberInRange;
 
-import racingcar.domain.game.GameNumberGenerator;
+import racingcar.domain.game.NumberGenerator;
 
-public class CanMoveGameNumberGenerator implements GameNumberGenerator {
+public class CanMoveNumberGenerator implements NumberGenerator {
     @Override
     public int generate() {
         return pickNumberInRange(4, 9);
